requirements.txt
typed-argument-parser
drake
pydot
ruff
pydantic
